Fixes/modifications for iSantePlus compatibility
iSantePlus (https://github.com/IsantePlus) is a OpenMRS 2.0.5
distribition used in Haiti. It extensively uses `Obs` to store a wide
range of medical data that might be stored in different part of the
OpenMRS data model in newer distributions, and that maps to multiple
FHIR resources. The FHIR2 module is being used for lab comminication and
continuity of care purposes. This commit contains fixes and
custom translators to enable the fhir2 module to be used with iSantePlus.

See [FM2-303](https://issues.openmrs.org/browse/FM2-303)