FM2-284: Add integration tests for DiagnosticReport
Obviously, this commit does a lot more than that :grin:.

I've refactored DiagnosticReport to use it's own table instead of being
backed by an Obs group. That makes it possible to add a DiagnosticReport
that refers to existing Obs and to update the DiagnosticReport when new
Obs are added.

Additionally, I added support for sorting by _lastUpdated, which *should*
generally work.

Finally, I noticed that we were returning URLs for ValueSets rather than
the coding systems that define the code behind those value sets. I've
changed this so that we always use the coding system URL, which is the
correct URL to use in this case.