(speculative) Always fetch the latest Context
It's been reported that, under some circumstances, the FHIR2 module being
restarted by OpenMRS's reload process results in no resource providers
being loaded. There are basically no steps to reproduce, so this is a
best-guess effort based on the idea that a context refresh might result
in the servlet hanging onto a Spring context with no defined beans in it.