Refactor LocationTranslatorImpl to add MFL code as an identifier
This commit refactors the LocationTranslatorImpl class to add the MFL (Master Facility List) code as an identifier to the FHIR Location resource. The MFL code is fetched from the location attribute defined in the global property MFL_LOCATION_ATTRIBUTE_TYPE_UUID. If the attribute is present, the MFL code is added as an identifier to the FHIR Location resource.